feat(RequestCore): Generate a new version of request core that uses KeywordManager
