+++ job_scrapper/scrap_job_from/inserm.py
@@ -3,7 +3,7 @@ import re
 from bs4 import BeautifulSoup
 
 import job_scrapper.scrapper_skeleton.pdf_jobs as pj
-from job_scrapper.scrapper_skeleton.request_core import ScrapperRequestCore
+from job_scrapper.scrapper_skeleton.legacy_request_core import ScrapperRequestCore
 
 
 class InsermScrapper(pj.PdfJobBaseScrapper):
